wip: add collections for {edge/node}-edits
- refactor: schema setup for DB, this is almost fully flechted migration
  code now!
